+++ src/test/java/org/springframework/samples/petclinic/ui/CreateAppointmentUITest.java
@@ -1,8 +1,6 @@
 package org.springframework.samples.petclinic.ui;
 
-import java.util.regex.Pattern;
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.util.concurrent.TimeUnit;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -21,7 +19,6 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class CreateAppointmentUITest {
     private WebDriver driver;
-    private String baseUrl;
     private boolean acceptNextAlert = true;
     private StringBuffer verificationErrors = new StringBuffer();
     private String appointmentDate = PetclinicDates.getFormattedFutureDate(LocalDate.now(), 5, "yyyy/MM/dd");
@@ -33,7 +30,6 @@ public class CreateAppointmentUITest {
     @BeforeEach
     public void setUp() throws Exception {
         driver = new FirefoxDriver();
-        baseUrl = "https://www.google.com/";
         driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
     }
 
@@ -46,7 +42,7 @@ public class CreateAppointmentUITest {
         driver.findElement(By.id("password")).clear();
         driver.findElement(By.id("password")).sendKeys("0wn3333r_1");
         driver.findElement(By.xpath("//button[@type='submit']")).click();
-        driver.findElement(By.xpath("//div[@id='main-navbar']/ul[2]/li/a/strong")).click();
+        driver.findElement(By.xpath("//div[@id='main-navbar']/ul[2]/li/a")).click();
         driver.findElement(By.linkText("My Profile")).click();
         driver.findElement(By.linkText("Add Appointment")).click();
         driver.findElement(By.xpath("//input[@id='appointmentDate']")).click();
@@ -58,9 +54,34 @@ public class CreateAppointmentUITest {
         new Select(driver.findElement(By.name("vet"))).selectByVisibleText("Rafael Ortega");
         driver.findElement(By.xpath("//option[@value='4']")).click();
         driver.findElement(By.xpath("//button[@type='submit']")).click();
-        assertEquals(expectedAppointmentDate, driver.findElement(By.xpath("//td[3]/table/tbody/tr/td")).getText());
+        assertEquals(expectedAppointmentDate, driver.findElement(By.xpath("//td[3]/table/tbody/tr[3]/td")).getText());
         assertEquals("Edit", driver.findElement(By.linkText("Edit")).getText());
         assertEquals("Delete", driver.findElement(By.linkText("Delete")).getText());
+    }
+    
+    @Test
+    public void testNotCreateAppointmentByPetUI() throws Exception {
+    	driver.get("http://localhost:" + port);
+        driver.findElement(By.xpath("//a[contains(text(),'Login')]")).click();
+        driver.findElement(By.id("username")).clear();
+        driver.findElement(By.id("username")).sendKeys("owner1");
+        driver.findElement(By.id("password")).clear();
+        driver.findElement(By.id("password")).sendKeys("0wn3333r_1");
+        driver.findElement(By.xpath("//button[@type='submit']")).click();
+        driver.findElement(By.xpath("//div[@id='main-navbar']/ul[2]/li/a")).click();
+        driver.findElement(By.linkText("My Profile")).click();
+        driver.findElement(By.linkText("Add Appointment")).click();
+        driver.findElement(By.xpath("//input[@id='appointmentDate']")).click();
+        driver.findElement(By.id("appointmentDate")).clear();
+        driver.findElement(By.id("appointmentDate")).sendKeys(appointmentDate);
+        driver.findElement(By.id("description")).click();
+        driver.findElement(By.id("description")).clear();
+        driver.findElement(By.id("description")).sendKeys("Malestar general");
+        driver.findElement(By.name("vet")).click();
+        new Select(driver.findElement(By.name("vet"))).selectByVisibleText("Rafael Ortega");
+        driver.findElement(By.xpath("//option[@value='4']")).click();
+        driver.findElement(By.xpath("//button[@type='submit']")).click();
+        assertEquals("Imposible realizar una cita con esos datos", driver.findElement(By.id("vetError")).getText());
     }
 
     @AfterEach
