+++ src/test/java/org/springframework/samples/petclinic/ui/DeleteAppointmentUITest.java
@@ -2,11 +2,13 @@ package org.springframework.samples.petclinic.ui;
 
 import java.util.regex.Pattern;
 import java.time.LocalDate;
+import java.util.List;
 import java.util.concurrent.TimeUnit;
 import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.hamcrest.CoreMatchers.*;
 import org.openqa.selenium.*;
 import org.openqa.selenium.firefox.FirefoxDriver;
@@ -18,12 +20,13 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@TestMethodOrder(OrderAnnotation.class)
 public class DeleteAppointmentUITest {
   private WebDriver driver;
-  private String baseUrl;
   private boolean acceptNextAlert = true;
   private StringBuffer verificationErrors = new StringBuffer();
-  private String appointmentDate = PetclinicDates.getFormattedFutureDate(LocalDate.now(), 5, "yyyy/MM/dd");
+  private String appointmentDateToDelete = PetclinicDates.getFormattedFutureDate(LocalDate.now(), 5, "yyyy/MM/dd");
+  private String appointmentDateToNotDelete = PetclinicDates.getFormattedFutureDate(LocalDate.now(), 1, "yyyy/MM/dd");
   
   @LocalServerPort
   private int port;
@@ -31,12 +34,12 @@ public class DeleteAppointmentUITest {
   @BeforeEach
   public void setUp() throws Exception {
     driver = new FirefoxDriver();
-    baseUrl = "https://www.google.com/";
     driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
   }
 
   @Test
-  public void testUntitledTestCase() throws Exception {
+  @Order(1)
+  public void testDeleteAppointment() throws Exception {
     driver.get("http://localhost:" + port);
     driver.findElement(By.xpath("//a[contains(@href, '/login')]")).click();
     driver.findElement(By.id("username")).clear();
@@ -49,13 +52,41 @@ public class DeleteAppointmentUITest {
     driver.findElement(By.linkText("Add Appointment")).click();
     driver.findElement(By.xpath("//input[@id='appointmentDate']")).click();
     driver.findElement(By.id("appointmentDate")).clear();
-    driver.findElement(By.id("appointmentDate")).sendKeys(appointmentDate);
+    driver.findElement(By.id("appointmentDate")).sendKeys(appointmentDateToDelete);
     driver.findElement(By.id("description")).clear();
     driver.findElement(By.id("description")).sendKeys("Description");
     driver.findElement(By.name("vet")).click();
     driver.findElement(By.xpath("//option[@value='1']")).click();
     driver.findElement(By.xpath("//button[@type='submit']")).click();
-    driver.findElement(By.linkText("Delete")).click();
+    int nAppointments = getNumberOfAppointments();
+    driver.findElement(By.xpath("//td[3]/table/tbody/tr["+ (nAppointments) +"]/td[4]/a")).click();
+    assertEquals(getNumberOfAppointments(), nAppointments - 1);
+  }
+  
+  @Test
+  @Order(2)
+  public void testNotDeleteAppointment() throws Exception {
+    driver.get("http://localhost:" + port);
+    driver.findElement(By.xpath("//a[contains(@href, '/login')]")).click();
+    driver.findElement(By.id("username")).clear();
+    driver.findElement(By.id("username")).sendKeys("owner1");
+    driver.findElement(By.id("password")).clear();
+    driver.findElement(By.id("password")).sendKeys("0wn3333r_1");
+    driver.findElement(By.xpath("//button[@type='submit']")).click();
+    driver.findElement(By.xpath("//a[contains(@href, '#')]")).click();
+    driver.findElement(By.xpath("//a[contains(@href, '/users/profile')]")).click();
+    driver.findElement(By.linkText("Add Appointment")).click();
+    driver.findElement(By.id("appointmentDate")).click();
+    driver.findElement(By.id("appointmentDate")).clear();
+    driver.findElement(By.id("appointmentDate")).sendKeys(appointmentDateToNotDelete);
+    driver.findElement(By.id("description")).clear();
+    driver.findElement(By.id("description")).sendKeys("Description");
+    driver.findElement(By.name("vet")).click();
+    driver.findElement(By.xpath("//option[@value='1']")).click();
+    driver.findElement(By.xpath("//button[@type='submit']")).click();
+    int nAppointments = getNumberOfAppointments();
+    driver.findElement(By.xpath("//td[3]/table/tbody/tr["+ (nAppointments) +"]/td[4]/a")).click();
+    assertEquals("No se puede cancelar una cita con dos o menos días de antelación", driver.findElement(By.className("error-text")).getText().trim());
   }
 
   @AfterEach
@@ -65,6 +96,10 @@ public class DeleteAppointmentUITest {
     if (!"".equals(verificationErrorString)) {
       fail(verificationErrorString);
     }
+  }
+  
+  private int getNumberOfAppointments( ) {
+	  return driver.findElements(By.xpath("//td[3]/table/tbody/tr")).size()-1;
   }
 
   private boolean isElementPresent(By by) {
