+++ src/test/java/org/springframework/samples/petclinic/ui/MedicalTestUITest.java
@@ -1,26 +1,31 @@
 package org.springframework.samples.petclinic.ui;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.concurrent.TimeUnit;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.openqa.selenium.Alert;
 import org.openqa.selenium.By;
 import org.openqa.selenium.NoAlertPresentException;
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class MedicalTestUITest {
 	private WebDriver driver;
 	private String baseUrl;
 	private boolean acceptNextAlert = true;
 	private StringBuffer verificationErrors = new StringBuffer();
+	
+    @LocalServerPort
+    private int port;
 
 	@BeforeEach
 	public void setUp() throws Exception {
@@ -31,7 +36,7 @@ public class MedicalTestUITest {
 
 	@Test
 	public void testAddMedicalTest() throws Exception {
-		driver.get("http://localhost:8080/");
+		driver.get("http://localhost:" + port);
 		driver.findElement(By.xpath("//div[@id='main-navbar']/ul[2]/li/a")).click();
 		driver.findElement(By.id("username")).click();
 		driver.findElement(By.id("username")).sendKeys("admin1");
@@ -58,7 +63,7 @@ public class MedicalTestUITest {
 
 	@Test
 	public void testAddMedicalTestWithErrors() throws Exception {
-		driver.get("http://localhost:8080/");
+		driver.get("http://localhost:" + port);
 		driver.findElement(By.xpath("//div[@id='main-navbar']/ul[2]/li/a")).click();
 		driver.findElement(By.id("username")).click();
 		driver.findElement(By.id("username")).clear();
